Update 06_leapYears README.md with a reference
I updated the README with an external link outlining the nested rules for leap years. The present problem statement, borrowed from Chris Pine, does not clearly state these rules, which is indeed confusing.
To scaffold the learner's journey, it would be useful to add a reference that walks them through the leap year rules with examples. It difficult to stick to the practice exercise if the learner has to shoot in the dark without knowing the exact rules of the game.