02_repeatString/solution - Line-32, replace "odin" with "hey"
As asked by @JoshDevHub in a discussion of a related PR, I made a similar change to this solution. In the earlier code, the comment description mentioned "hey" as a sample string but the code had "odin" string. This update addresses that tiny gap between sample comments and related code.

More on: https://github.com/TheOdinProject/javascript-exercises/pull/505#pullrequestreview-2588692590